rename clmm_cosmo, clmm_modeling -> parent_class
